Add forcast detail scroll as a subclass of Composite
Use more up-to-date type definitions for Tabris.js to refactor the forcast detail scroll as a subclass of the Composite widget.
Display a background image and an instance of the new forecast scroll .
Also added an activity indicator while polling for weather information.
Some changes to weatherServices.ts were errorneously not commitet in previous commit.